Reescreve issue_alerts.py para coletar dados direto do website SciELO Brasil, eliminando dependência do Articlemeta e Xylose Refs #34
