+++ issue_alerts/issue_alerts.py
@@ -1,16 +1,13 @@
-# coding: utf-8
 import configparser
 import datetime
-import json
 import logging
 import os
+import re
 import sys
 
 import requests
 from bs4 import BeautifulSoup
-from bs4 import Comment
 from jinja2 import Environment, FileSystemLoader, Template
-from xylose.scielodocument import Article, Issue
 
 
 # Check and create logs directorycd pro
@@ -32,204 +29,133 @@ config.read('config.ini')
 invalid_sec = config['invalid_sec']['sections'].encode('latin-1').decode('utf-8').split(', ')
 invalid_sec = invalid_sec + [s.upper() for s in invalid_sec]
 
+# Domain
+domain = 'https://www.scielo.br'
 
-def leave():
+# Set Language EN from WebSite in new Session
+try:
+    print('Starting a website session...')
+    s = requests.Session()
+    s.get("https://www.scielo.br/set_locale/en/")
+except Exception as e:
+    print(e)
+    sys.exit()
+    
+
+def leave(e=None):
+    if e:
+        print(e)
     i = input("\nPress a key to exit... ")
     if i != '':
         sys.exit()
 
 
-def getpidcode(url):
-    try:
-        r = requests.get(url)
+def remove_ilang(url):
+    return re.sub(r'([&?])ilang=[^&]*&?', r'\1', url).rstrip('&?')
+
+
+def getdocs(urli, articles):
+    # Get Issue page
+    # urli = "https://www.scielo.br/j/ean/i/2021.v25n1/"
+    # urla = "https://www.scielo.br/j/pg/a/HrVWp3P85qGksXDqQM9ZdSx/?lang=de"
+
+    # Issue
+    if urli and urli != None:
+        
+        r = s.get(urli, cookies=s.cookies)
         soup = BeautifulSoup(r.content, "html.parser")
+        docs = soup.find_all("td", {"class":"pt-4"})
+        
+        print('Total documents: %s\n' % (len(docs)))
 
-        # get PIDs
-        pidscomm = soup.find_all(string=lambda text: isinstance(text, Comment) and 'PID:' in text)
-        pids = [pid.strip(' PID: ') for pid in pidscomm]
+        return docs
+    
+    # Articles
+    if articles and articles != None:
+        
+        docsa = []
+        lurli = None
 
-        # get Codes (new PID) and Prefix
-        if url.split('/')[5] == 'a':
-            # is article
-            codes = [url.split('/a/')[1].split('/')[0]]
-            prefix = [url.split('/a/')[0] for p in range(len(pids))]
-        else:
-            # is issue
-            arts = soup.find_all("td", attrs={"data-date":True})
-            codes = [art.find_all('a')[-1].attrs['href'].split('/')[4] for art in arts]
-            prefix = [url.split('/i/')[0] for p in range(len(pids))]
-
-        # zip Prefix, Codes and PIDs
-        pidscodelist = list(zip(prefix, codes, pids))
-
-        return pidscodelist
-
-    except Exception as e:
-        print(e)
-        leave()
-
-
-def requestissue(config, issue_pid):
-    # Request Issue
-    # https://articlemeta.scielo.org/api/v1/issue/?code=0104-070720190001
-    uissue = config['articlemeta']['host']+'/api/v1/issue/?code=%s' % issue_pid
-    logger.info(uissue)
-
-    xissue = None
-    while xissue is None:
-        try:
-            rissue = requests.get(uissue)
-            xissue = Issue(rissue.json())
-        except requests.exceptions.Timeout:
-            logger.info('error: %s' % e)
-            print("Timeout - Try again")
-            leave()
-        except requests.exceptions.RequestException as e:
-            logger.info('error: %s' % e)
-            print("Request Error - Check your connection and try again")
-            leave()
-        except json.decoder.JSONDecodeError as e:
-            logger.info('error: %s' % e)
-            print("Request Error - Try again")
-            leave()
-
-    # Valid Codes list
-    seccode_list = []
-    if xissue.sections != None:
-        for i, sec in enumerate(list(xissue.sections.values())):
-             for sectext in list(sec.values()):
-                 if sectext not in invalid_sec and list(xissue.sections)[i] not in seccode_list:
-                     seccode_list.append(list(xissue.sections)[i])
-
-    return (xissue, seccode_list)
+        for urla in articles:
+            print(urla)
+            # get issue page
+            pidv3 = urla.split('/')[6]
+            r = s.get(urla, cookies=s.cookies)
+            soup = BeautifulSoup(r.content, "html.parser")
+            tc = soup.find_all("a", {"class":"btn mb-0"})[0]['href']
+            urli = domain + tc
+            
+            # access Table of contents of the article
+            if urli != lurli:
+                r = s.get(urli, cookies=s.cookies)
+                soup = BeautifulSoup(r.content, "html.parser")
+                docsi = soup.find_all("td", {"class":"pt-4"})
+                
+            # insert 1 doc in docs list
+            for doc in docsi:
+                for link in doc.find_all("a", {"class":"text-uppercase"}):
+                    if pidv3 in link.attrs['href']:
+                        docsa.append(doc)
+                        break
+            
+            # preserves urli
+            lurli = urli
+        
+        if docsa:
+            print('Total documents: %s\n' % (len(docsa)))
+            
+            return docsa
 
 
 def json2html(htmlout, config, urli=None, articles=None):
-    # get PID and Codes
-    print('Getting list of pids...')
-    if urli:
-        pid_code_list = getpidcode(urli)
-
-    if articles:
-        pid_code_list = [getpidcode(urla)[0] for urla in articles]
-
-    print('Total documents: %s\n' % (len(pid_code_list)))
 
     # Write the HTML file
     with open(htmlout, encoding='utf-8', mode='w') as f:
         # Start HTML output
         f.write(u'<html>\n<body>\n')
 
-        # Request Issue
-        reqissue = requestissue(config, pid_code_list[0][2][1:18])
-        xissue = reqissue[0]
-        seccode_list = reqissue[1]
-
         # JINJA
         jinja_env = Environment(loader=FileSystemLoader('template'))
         template = jinja_env.get_template('body.html')
 
+        # First section only
         previous_sec = None
+        section = None
 
-        for prefix, code, pid in pid_code_list:
-            
-            # Request Article
-            uart = config['articlemeta']['host']+"/api/v1/article/?code=%s" % pid
-            xart = None
-            while xart is None:
-                try:
-                    rart = requests.get(uart)
-                    xart = Article(rart.json())
-                except requests.exceptions.Timeout:
-                    logger.info('error: %s' % e)
-                    print("Timeout - Try again")
-                    leave()
-                except requests.exceptions.RequestException as e:
-                    logger.info('error: %s' % e)
-                    print("Request Error - Check your connection and try again")
-                    leave()
-                except json.decoder.JSONDecodeError as e:
-                    logger.info('error: %s' % e)
-                    print("Request Error - Try again")
-                    leave()
+        # Get a list of docs from Issue or a Article
+        docs = getdocs(urli, articles)
 
-            # Language priority to HTML
-            lang_priority = ['en', 'pt', 'es', 'fr', 'it', 'de', 'ru']
+        for doc in docs:
 
-            # Sets the language of the template
-            for l in lang_priority:
-                if l in xart.languages():
-                    lang = l
-                    break
-
-            # First section only
-            if xart.section_code:
-                if xart.section_code in xissue.sections.keys():
-                    if 'en' in xissue.sections[xart.section_code].keys():
-                        section = xissue.sections[xart.section_code]['en'].upper()
-                    elif lang in xissue.sections[xart.section_code].keys():
-                        section = xissue.sections[xart.section_code][lang].upper()
-            else:
+            # SECTION
+            try:
+                # previous_sec = doc.find("span", {"class":"badge"}).text
+                section = doc.find("span", {"class":"badge"}).text.upper()
+            except:
+                # previous_sec = "ORIGINAL ARTICLE"
                 section = "ORIGINAL ARTICLE"
 
             if section:
-                if previous_sec != section and section.upper() not in invalid_sec:
-                    print(section)
+                if section != previous_sec and section.upper() not in invalid_sec:
+                    print('\n' + section)
                     tsec = Template("<p><strong>{{ section }}</strong></p>\n\n")
                     outsec = tsec.render(section=section)
                     f.write(outsec)
                     previous_sec = section
-            else:
-                logger.info('Section Error: %s' % pid)
-                print('Section Error: %s' % pid)
-
-            # Article metadata
-            if section:
+                      
                 if section.upper() not in invalid_sec:
                     # Title
                     title_html = None
                     title = None
-                    # Scraping HTML title
-                    try:
-                        # prioritizes english language
-                        link = ('%s/a/%s/?lang=en' % (prefix, code))
-                        r = requests.get(link)
-                        soup = BeautifulSoup(r.content, 'html.parser')
-                        arttitle = soup.find("h1", {"class":"article-title"})
-                        
-                        # Clear tags and attributes
-                        [img.decompose() for img in arttitle.find_all('img')]
-                        [a.decompose() for a in arttitle.find_all('a')]
-                        [st.decompose() for st in arttitle.find_all('strong')]
-                        [sp.decompose() for sp in arttitle.find_all('span')]
-                        arttitle.attrs.clear()
-                        arttitle.name = 'strong'
-                        title_html = arttitle
-                    except requests.exceptions.Timeout:
-                        logger.info('error: %s' % e)
-                        print("Timeout - Try again")
-                        leave()
 
-                    ## HTML title or original_title
-                    if title_html:
-                        title = title_html
-                    elif xart.original_language() == lang:
-                        title = xart.original_title()
-                    elif lang in xart.translated_titles().keys():
-                        title = xart.translated_titles()[lang]
-                    else:
-                        title = xart.original_title()
-
-                    # show PID title to user
-                    if title_html:
-                        print(pid, title.text.strip()[0:60])
-                    else:
-                        print(pid, title.strip()[0:60])
-
+                    # Scraping HTML
+                    title_html = doc.find("strong", {"class":"d-block mt-2"})
+                    title_html.attrs.clear()
+                    title = title_html.text
+                    
                     # Authors
                     authors = []
-                    if xart.authors:
-                        authors = [au['surname']+', '+au['given_names'] for au in xart.authors]
+                    authors = [au.text for au in doc.find_all("a", {"class":"me-2"})]
 
                     # Link text in english
                     link_text = {
@@ -241,62 +167,56 @@ def json2html(htmlout, config, urli=None, articles=None):
                             'de': ('text in German', 'German'),
                             'ru': ('text in Russian', 'Russian'),
                             }
+                    
+                    # ALL Links
+                    all_links = [link.attrs['href'] for link in doc.find_all("a", {"class":"text-uppercase"})]
 
-                    # Valid languages dict
-                    vld = [
-                        {'en': 'Inglês'}, 
-                        {'en': 'English'}, 
-                        {'en': 'Inglés'},
-                        {'pt': 'Português'},
-                        {'pt': 'Portuguese'},
-                        {'pt': 'Portugués'},
-                        {'es': 'Espanhol'},
-                        {'es': 'Spanish'},
-                        {'es': 'Espanõl'},
-                        {'fr': 'Français'},
-                        {'it': 'Italiano'},
-                        {'de': 'Deutsch'},
-                        {'ru': 'Русский'}] 
+                    # PID v3
+                    for a in all_links:
+                        if 'format=pdf' in a:
+                            pidv3 = a.split('/')[4]
+                            break
+                    
+                    # show PID v3 title to user   
+                    print(pidv3, title.strip()[0:60])
                     
                     # Text Links
-                    llinktxt = None
-                    btntxt = soup.find_all("ul", {"aria-labelledby": "btnGroupDropTextLanguage"})[0]
-                    llinktxt = btntxt.find_all('a')
-
-                    langtxt = []
-                    for a in llinktxt:
-                        for d in vld:
-                            for k, v in d.items():
-                                if v in a.text.strip():
-                                    langtxt.append(k)
-
-                    if llinktxt:
-                        ltxt = []
-                        for l in langtxt:
-                            utxt = '%s/a/%s/?lang=%s' % (prefix, code, l)
+                    ltxt = []
+                    for a in all_links:
+                        if 'abstract' not in a and 'format=pdf' not in a:
+                            # remove ilang 
+                            a_clean = remove_ilang(a)
+                            # get language 
+                            l = a_clean[-2:]
+                            # article url
+                            utxt = '%s%s' % (domain, a_clean)
                             ltxt.append((link_text[l][0], link_text[l][1], utxt))
                     
                     # PDF Links
-                    llinkpdf = None
-                    btnpdf = soup.find_all("ul", {"aria-labelledby": "btnGroupDropPDF"})[0]
-                    llinkpdf = btnpdf.find_all('a')
-
-                    if llinkpdf:
-                        lpdf = []
-                        for l in langtxt: # uses languagues from text links
-                            updf = '%s/a/%s/?format=pdf&lang=%s' % (prefix, code, l)
+                    lpdf = []
+                    for a in all_links:
+                        if 'format=pdf' in a:
+                            # remove ilang 
+                            a_clean = remove_ilang(a)
+                            # get language 
+                            l = a_clean[-2:]
+                            # pdf url
+                            updf = '%s%s' % (domain, a_clean)
                             lpdf.append((link_text[l][1], updf))
 
                     # Render HTML
                     output = template.render(
-                        title=title, authors=authors, lpdf=lpdf, ltxt=ltxt)
+                        title=title_html, authors=authors, lpdf=lpdf, ltxt=ltxt)
+                    
                     f.write(output)
 
         # Terminate HTML output
         f.write(u'</body>\n</html>')
+        f.close()
 
 
 def main():
+    
     # Folder and file names
     if config['paths']['issuelistname'] == '':
         print('issuelistname = empty.\nEnter a name for the issue list in config.ini.')
@@ -325,31 +245,33 @@ def main():
     if not os.path.exists(htmlfolder):
         os.mkdir(htmlfolder)
 
-    # ISSUE or Article List
+    # ISSUE or ARTICLE List
     with open(config['paths']['issuelistname']) as f:
         urllist = [line.strip() for line in f]
     f.close()
-
-    articles = []
-    for url in urllist:
-        if url.split('/')[5] == 'a':
-            articles.append(url)
 
     issues = []
     for url in urllist:
         if url.split('/')[5] == 'i':
             issues.append(url)
 
+    articles = []
+    for url in urllist:
+        if url.split('/')[5] == 'a':
+            articles.append(url)
+
     if issues:
-        for urli in issues:
-            issue = urli.split('/')[4]+'_'+urli.split('/')[6]
+        for url in issues:
+            issue = url.split('/')[4]+'_'+url.split('/')[6]
             htmlout = ('%s/%s_%s.html' % (htmlfolder, htmlfilename, issue))
             print('\nfolder/htmlfile: %s\n' % htmlout)
             
             # Build HTML object
-            json2html(htmlout=htmlout, config=config, urli=urli, articles=None)
+            print(url)
+            json2html(htmlout=htmlout, config=config, urli=url, articles=None)
 
     if articles:
+        # for url in articles:
         acron = articles[0].split('/')[4]
         htmlout = ('%s/%s_%s_articles.html' % (htmlfolder, htmlfilename, acron))
         print('\nfolder/htmlfile: %s\n' % htmlout)
