+++ issue_alerts/requirements.txt
@@ -1,5 +1,4 @@
 Jinja2==3.1.6
 requests==2.32.2
-xylose==1.35.4
 PyInstaller==6.10.0
 beautifulsoup4==4.13.3
